Moved and renamed some header files
