+++ libraries/deltaRobot/include/DeltaRobot/EffectorBoundariesException.h
@@ -47,7 +47,7 @@
 #include <sstream>
 
 
-namespace huniplacer
+namespace DeltaRobot
 {
